+++ expense-manage/src/main/java/com/porejct/expensemanage/commone/security/handler/AuthenticationFailureCustomHandler.java
@@ -1,10 +1,6 @@
 package com.porejct.expensemanage.commone.security.handler;
 
-import com.porejct.expensemanage.commone.exception.ErrorResponse;
-import com.porejct.expensemanage.commone.security.exception.AuthExceptionCode;
-import com.porejct.expensemanage.commone.utils.response.ResponseUtils;
-import com.porejct.expensemanage.commone.utils.translator.ObjectMapperUtils;
-import jakarta.servlet.ServletException;
+import com.porejct.expensemanage.commone.utils.response.ErrorResponseUtils;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
@@ -18,26 +14,16 @@ import org.springframework.stereotype.Component;
 @Component
 @RequiredArgsConstructor
 public class AuthenticationFailureCustomHandler implements AuthenticationFailureHandler {
-    private final ObjectMapperUtils objectMapperUtils;
+
+    private final ErrorResponseUtils errorResponseUtils;
 
     @Override
     public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response,
             AuthenticationException exception) throws IOException {
         log.error("# Authentication failed: {}", exception.getMessage());
         log.error("authentication ", exception);
-        getSendErrorResponse(response);
+
+        errorResponseUtils.sendErrorResponse(response);
     }
 
-    private void getSendErrorResponse(HttpServletResponse response) throws IOException {
-        new ResponseUtils(response,getResponseData()).sendErrorResponse();
-    }
-
-    private String getResponseData() {
-        return objectMapperUtils.toStringValue(createErrorResponse());
-    }
-
-    private ErrorResponse createErrorResponse() {
-        return ErrorResponse.of(AuthExceptionCode.UNAUTHENTICATED.getHttpStatus(),
-                AuthExceptionCode.UNAUTHENTICATED.getMessage());
-    }
 }
